Refactor code
WHAT:
- Move fuctions into Utils file to reuse in other
files
- Refactor some sample data
